+++ bigjobasync/output_transfer_worker.py
@@ -55,6 +55,8 @@ class _OutputTransferWorker(multiprocessing.Process):
                     # 'ready_to_exec' or 'failed' queues. 
                     self.transfer_output_file(task)
                     #self._tasks_ready_to_exec_q.put(task)
+                    self._tasks_ready_to_transfer_output_q.task_done()
+
                 except Queue.Empty:
                     break