+++ examples/02_example_local_remote_input.py
@@ -98,7 +98,9 @@ if __name__ == "__main__":
     #    * FAILED          (an error occured during pilot execution)
     #
     stampede.register_callbacks(resource_cb)
-    stampede.allocate()
+    # If terminate_on_empty_queue=True, the resource will be shut down as soon
+    # as the last task has finished. 
+    stampede.allocate(terminate_on_empty_queue=True)
 
     # Define tasks and their input and output files
     all_tasks = []