+++ CHANGES.md
@@ -1,3 +1,12 @@
+Version 0.2
+----------------------------------------------------------------------
+
+* Improved examples
+* Added sample data
+* Added support for resource.allocate(terminate_on_empty_queue=True)
+  This flag terminates the BigJob once there are no more pending jobs.
+
+
 Version 0.1
 ----------------------------------------------------------------------